+++ example/otel-collector/main.go
@@ -177,6 +177,10 @@ func main() {
 func connectDB() (*sql.DB, func()) {
 	attrs := append(otelsql.AttributesFromDSN(mysqlDSN), semconv.DBSystemNameMySQL)
 
+	if dbNamespace, err := otelsql.DBNamespaceFromDSN(mysqlDSN); err == nil {
+		attrs = append(attrs, dbNamespace)
+	}
+
 	// Connect to database
 	db, err := otelsql.Open("mysql", mysqlDSN, otelsql.WithAttributes(
 		attrs...,